refactor into general policy data point superclass
